10/05/2026

+++ going_modular/make_columns_object.py
@@ -1,6 +1,6 @@
 def make_columns_object(df, cols):
     for col in cols:
-        df[col] = df[col].astype('object')
+        df[col] = df[col].astype('object').astype(int)
     return df
 
 # df = make_columns_object(df, ['European', 'Event'])
